OAKJAVA7
OAKJAVA7(JAVA7.0) Programming Language is  invented by  wilmix jemin  j  at  year  2013,2016 in  java ,DOTWEB3.0. and OAKJAVA7(JAVA7.0)  is  a   World No:2 Programming Language and  it  is   focused  on   creating  Prototype  software,used  to  create Gaming and utilities,  used  in  web application  with  java frameworks , used  with  remote   webapplication  with  java frameworks,  and  it  has  Advanced Oops. It  also  supports obfuscation, Code Security, and it  is Learnable .